% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_formula_with_species_interaction.R
\name{build_formula_with_species_interaction}
\alias{build_formula_with_species_interaction}
\title{Build Formula with Species Interaction}
\usage{
build_formula_with_species_interaction(
  rhs,
  response = "y",
  species_var = "OTU",
  lower_level = "Replicate"
)
}
\arguments{
\item{rhs}{The right-hand side of the formula as a `formula` object (e.g., `~ (1 | Site) + treatment`).}

\item{response}{A character string specifying the response variable (default: `"y"`).}

\item{species_var}{The name of the species variable to interact with each term (default: `"OTU"`).}

\item{lower_level}{A variable name (usually a random effect) that should NOT be interacted with the species (default: `"Replicate"`).}
}
\value{
A `formula` object with the specified response and species interaction terms.
}
\description{
Automatically constructs a model formula where each covariate term 
(except a designated lower-level random effect) is interacted with 
the species variable. This is used in occupancy-abundance modeling 
to fit multi-species GLMMs.
}
\examples{
build_formula_with_species_interaction(
  rhs = ~ (1 | Site) + (1 | Sample) + (1 | Replicate) + treatment,
  response = "y",
  species_var = "OTU",
  lower_level = "Replicate"
)

}
