% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models_cv_gllvm.R
\name{compare_models_cv_gllvm}
\alias{compare_models_cv_gllvm}
\title{Cross-validation model comparison for FitModel_gllvm}
\usage{
compare_models_cv_gllvm(
  phyloseq,
  models,
  site_col,
  sample_col,
  replicate_col = NULL,
  otu_col,
  count_col,
  cv_level = c("site", "sample", "OTU", "replicate"),
  K = 5,
  seed = 123
)
}
\arguments{
\item{phyloseq}{A phyloseq object.}

\item{models}{Named list of model-fitting functions.}

\item{site_col}{Character. Site column name.}

\item{sample_col}{Character. Sample column name.}

\item{replicate_col}{Character. Replicate column name.}

\item{otu_col}{Character. OTU column name.}

\item{count_col}{Character. Count column name.}

\item{cv_level}{Character. One of:
\code{"site"}, \code{"sample"},
\code{"OTU"}, or \code{"replicate"}.}

\item{K}{Integer. Number of folds.}

\item{seed}{Integer random seed.}
}
\value{
A named list containing fold-by-fold
cross-validation metrics for each model.
}
\description{
Performs K-fold cross-validation for hierarchical occupancy–
capture–abundance models fitted using \code{FitModel_gllvm()}.
}
\details{
The function evaluates predictive performance using:
\itemize{
  \item Occupancy AUC (\code{psi_AUC})
  \item Capture AUC (\code{capture_AUC})
  \item Predictive abundance log-likelihood
        (\code{lambda_log_score})
}
}
\examples{
\dontrun{

cv_res <- compare_models_cv_gllvm(
  phyloseq = ps,
  models = models,
  site_col = "site_month",
  sample_col = "Name",
  replicate_col = "Replicate",
  otu_col = "OTU",
  count_col = "y",
  cv_level = "sample",
  K = 3
)

}
}
