% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_metrics.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute predictive performance metrics}
\usage{
compute_metrics(obs_df, pred_df, metric = "log_score", count_col = "y")
}
\arguments{
\item{obs_df}{Observed data (must contain count column)}

\item{pred_df}{Predictions (must contain lambda_mean)}

\item{metric}{Metric name: "log_score" or "rmse"}

\item{count_col}{Name of count column}
}
\value{
Numeric score
}
\description{
Compute predictive performance metrics
}
