% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_function.R
\name{compute_residuals_TMB}
\alias{compute_residuals_TMB}
\title{Compute Residuals from TMB Model}
\usage{
compute_residuals_TMB(model, data_array_filtered)
}
\arguments{
\item{model}{Output from run_full_TMB}

\item{data_array_filtered}{Original input array (Species x Sites x Replicates)}
}
\value{
A list with residual matrices: occupancy_residuals, abundance_residuals
}
\description{
Computes Pearson residuals for both occupancy and abundance components.
}
