% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylosq_function_documented.R
\name{data_array_phyloseq}
\alias{data_array_phyloseq}
\title{Generate Data Array from Phyloseq Object}
\usage{
data_array_phyloseq(phyloseq_path, verbose = TRUE)
}
\arguments{
\item{phyloseq_path}{Path to the .RDS file containing a phyloseq object.}

\item{verbose}{Logical, if TRUE (default), prints progress messages.}
}
\value{
A 3D array with dimensions (Species x Sites x Replicates).
}
\description{
This function reads a phyloseq object and converts it into a 3D array (Species x Sites x Replicates).
}
\examples{
\dontrun{
data_array <- data_array_phyloseq("https://github.com/marwah83/eDNAModel/blob/main/Data/longdataexample.RDS")
}

}
