% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDNAModel-package.R
\docType{package}
\name{eDNAModel-package}
\alias{eDNAModel}
\alias{eDNAModel-package}
\title{eDNAModel}
\description{
Tools for fitting occupancy detection models for environmental DNA data.
}
\author{
\strong{Maintainer}: Marwah Soliman \email{marwah.soliman@ntnu.no}

Authors:
\itemize{
  \item Marwah Soliman \email{marwah.soliman@ntnu.no}
  \item Bert van der Veen \email{Bert.van-der-Veen@uni-bayreuth.de}
}

}
\keyword{internal}
