% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_cv_gllvm.R
\name{evaluate_cv_gllvm}
\alias{evaluate_cv_gllvm}
\title{Evaluate cross-validation results}
\usage{
evaluate_cv_gllvm(cv_res)
}
\arguments{
\item{cv_res}{A list returned by
\code{compare_models_cv_gllvm()} or
\code{compare_models_cv()}.}
}
\value{
A data.frame summarising model performance.
}
\description{
Summarises cross-validation results from
\code{compare_models_cv_gllvm()} or
\code{compare_models_cv()}.
}
\details{
Computes:
\itemize{
  \item Mean occupancy AUC
  \item Mean capture AUC
  \item Mean predictive abundance log-likelihood
}

Models are ranked by predictive abundance
log-likelihood (\code{lambda_log_score}),
where larger values are better.
}
\examples{
\dontrun{

results_table <- evaluate_cv_gllvm(cv_res)

print(results_table)

}
}
