% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_physeq_new.R
\docType{data}
\name{physeq_new}
\alias{physeq_new}
\title{Example phyloseq dataset for eDNAModel}
\format{
A list of phyloseq objects.
}
\source{
Environmental DNA monitoring study (example dataset)
}
\usage{
physeq_new
}
\description{
A phyloseq dataset containing eDNA sequencing data used to demonstrate
the occupancy–detection modeling workflow implemented in the
`eDNAModel` package.
}
\details{
The object is a list of phyloseq objects corresponding to different
sampling locations. Each element contains OTU tables, taxonomic
information, and sample metadata used for model fitting and
visualization.


The dataset includes marine invasive species monitoring data from
Thessaloniki. It is used in the package vignette to illustrate
the occupancy–detection modeling workflow.
}
\examples{
data(physeq_new)
names(physeq_new)

}
\keyword{datasets}
