% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_predict.R
\name{predict_FitModel}
\alias{predict_FitModel}
\title{Predict from FitModel object}
\usage{
predict_FitModel(fit, newdata)
}
\arguments{
\item{fit}{Output from FitModel}

\item{newdata}{Data frame for prediction}
}
\value{
Data frame with predictions
}
\description{
Predict from FitModel object
}
