% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_summary.R
\name{print.summary.eDNAModel}
\alias{print.summary.eDNAModel}
\title{Print Method for summary.eDNAModel Objects}
\usage{
\method{print}{summary.eDNAModel}(x, ...)
}
\arguments{
\item{x}{A \code{summary.eDNAModel} object returned by \code{\link{summary.eDNAModel}}.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Custom print method for summary results from \code{eDNAModel}.
}
\seealso{
\code{\link{summary.eDNAModel}}
}
