% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_full_TMB_pipeline.R
\name{run_full_TMB}
\alias{run_full_TMB}
\title{Full TMB Pipeline with Results Extraction (Complete & Preserved Code)}
\usage{
run_full_TMB(data_array_filtered)
}
\arguments{
\item{data_array_filtered}{Filtered 3D data array ready for TMB model fitting.}
}
\value{
List containing optimized object `opt2` and final extracted results (occupancy, lambda, prob.detect)
}
\description{
This function runs a full TMB model fitting pipeline and returns occupancy, lambda, and detection probabilities.
}
