% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_utils.R
\name{subset_phyloseq_by_sites}
\alias{subset_phyloseq_by_sites}
\title{Subset phyloseq object by site}
\usage{
subset_phyloseq_by_sites(physeq, site_col, keep_sites)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{site_col}{Column name in sample_data defining sites}

\item{keep_sites}{Vector of site names to retain}
}
\value{
A subsetted phyloseq object
}
\description{
Subset phyloseq object by site
}
