% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylosq_function_documented.R
\name{data_array_phyloseq}
\alias{data_array_phyloseq}
\title{Generate Data Array from Phyloseq Object}
\usage{
data_array_phyloseq(phyloseq_path, verbose = TRUE)
}
\arguments{
\item{phyloseq_path}{Path to the .RDS file containing a phyloseq object.}

\item{verbose}{Logical, if TRUE (default), prints progress messages.}
}
\value{
A 3D array with dimensions (Species x Sites x Replicates).
}
\description{
Converts a phyloseq object into a 3D data array (Species x Sites x Replicates).
}
\examples{
data_path <- system.file("extdata", "longdataexample.RDS", package = "eDNAModel")
data_array <- data_array_phyloseq(data_path)

}
