% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitModel.R
\name{FitModel}
\alias{FitModel}
\title{Fit Hierarchical Occupancy-Detection Model}
\usage{
FitModel(
  phyloseq,
  site_col,
  abundance_rhs,
  occupancy_rhs,
  min_species_sum = 50,
  abundance_threshold = 1,
  n_iter = 50,
  burn_in = 10,
  abundance_family = "poisson"
)
}
\arguments{
\item{phyloseq}{A phyloseq object with OTU table and sample metadata.}

\item{site_col}{Column name for site grouping.}

\item{abundance_rhs}{Formula for abundance model (e.g. (1 | OTU) + ...).}

\item{occupancy_rhs}{Formula for occupancy model.}

\item{min_species_sum}{Filter OTUs by total abundance.}

\item{abundance_threshold}{Detection threshold.}

\item{n_iter}{Number of iterations.}

\item{burn_in}{Number of burn-in iterations.}

\item{abundance_family}{GLMM family for abundance model ("poisson", "nbinom", "zip", "zinb").}
}
\value{
A list with model fits and summaries.
}
\description{
This function fits occupancy-detection models using GLMMs.
}
