% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitModel.R
\name{FitModel}
\alias{FitModel}
\title{Fit Multi-Species Occupancy-Abundance Model from a phyloseq Object}
\usage{
FitModel(
  phyloseq,
  poisson_rhs,
  binomial_rhs,
  min_species_sum = 50,
  sampletype_keep = NULL,
  num_iterations = 100,
  burn_in = 50
)
}
\arguments{
\item{phyloseq}{A `phyloseq` object containing OTU, taxonomy, and sample metadata.}

\item{poisson_rhs}{A `formula` object defining the right-hand side of the Poisson model.}

\item{binomial_rhs}{A `formula` object defining the right-hand side of the Binomial model.}

\item{min_species_sum}{Minimum total count required to retain a taxon (default: 50).}

\item{sampletype_keep}{Optional character value. If specified, filters the data to keep only samples with this `sampletype`.}

\item{num_iterations}{Total number of iterations to run the model fitting (default: 100).}

\item{burn_in}{Number of iterations to discard as burn-in (default: 50).}
}
\value{
A list with fitted Poisson and Binomial model objects for each post-burn-in iteration.
}
\description{
Prepares long-format data from a `phyloseq` object, builds species-interacted
GLMM formulas, and fits Poisson (abundance) and Binomial (occupancy) models
using an iterative burn-in approach.
}
\examples{
\dontrun{
data(physeq_example)
result <- FitModel(
  phyloseq = physeq_example,
  poisson_rhs = ~ (1 | Site) + (1 | Sample) + (1 | Replicate) + treatment,
  binomial_rhs = ~ (1 | Site) + treatment,
  sampletype_keep = "biologicalsample"
)
}

}
