% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction function.R
\name{predict_TMB}
\alias{predict_TMB}
\title{Predict from TMB Model with Confidence Intervals (Safe Version)}
\usage{
predict_TMB(model, newX, formula,
            which = c("abundance", "occupancy"),
            type = c("response", "link", "terms"),
            level = 0.95, se = TRUE)
}

\arguments{
\item{model}{Output from run_full_TMB()}

\item{newX}{New data.frame of covariates}

\item{which}{One of "abundance" or "occupancy".}

\item{type}{One of "response", "link", or "terms".}

\item{level}{Confidence level for intervals.}

\item{se}{Logical. Return standard errors?}

\item{formula}{The same formula used for abundance or occupancy}

}
\value{
Data frame with predictions and CIs
}
\description{
Predict from TMB Model with Confidence Intervals (Safe Version)
}
