% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_full_TMB.R
\name{run_full_TMB}
\alias{run_full_TMB}
\title{Run Full TMB Model}
\usage{
run_full_TMB(
  data_array_filtered,
  covariate_data,
  a.formula = ~site + (1 | replicate),
  o.formula = ~site
)
}
\arguments{
\item{data_array_filtered}{A 3D array (Species x Sites x Replicates) after filtering.}

\item{covariate_data}{A data frame containing covariates for model matrices.}

\item{a.formula}{A formula for the abundance model. Default is \code{~ site + (1 | replicate)}.}

\item{o.formula}{A formula for the occupancy model. Default is \code{~ site}.}
}
\value{
A list containing the fitted TMB object and optimization result.
}
\description{
Fits a hierarchical multi-species occupancy-abundance model using TMB,
with support for covariate formulas and random effects.
}
