% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{summary.eDNAModel}
\alias{summary.eDNAModel}
\alias{summary}
\title{Summary of TMB Model Coefficients}
\usage{
\method{summary}{eDNAModel}(object, ...)
}
\arguments{
\item{object}{Output from \code{\link{run_full_TMB}} (an object of class \code{eDNAModel})}

\item{...}{Additional arguments passed to TMB's summary function.}

\item{getJointPrecision}{\code{logical}, defaults to FALSE.}
}
\value{
A data frame with parameter names, estimates, and standard deviations.
}
\description{
Summarizes fixed-effect and random-effect parameters with standard deviations.
}
