% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{summary.eDNAModel}
\alias{summary.eDNAModel}
\alias{summaryall}
\title{Summary of TMB Model Coefficients}
\usage{
\method{summary}{eDNAModel}(model)
}
\arguments{
\item{model}{Output from \code{\link{run_full_TMB}}}
}
\value{
Data frame with parameter names, estimates, and standard deviations
}
\description{
Summarizes fixed-effect and random-effect parameters with standard deviations.
}
